fix(Modulo-1 Golpe): Alterar fatorDano de int para double

+++ CDZ-Modulo-1/Golpe.java
@@ -1,9 +1,9 @@
 
 public class Golpe {
     private String nome;
-    private int fatorDano;
+    private double fatorDano;
 
-    public Golpe(String nome, int fatorDano) {
+    public Golpe(String nome, double fatorDano) {
         this.nome = nome;
         this.fatorDano = fatorDano;
     }
@@ -12,7 +12,7 @@ public class Golpe {
         return nome;
     }
 
-    public int getFatorDano() {
+    public double getFatorDano() {
         return fatorDano;
     }
